requirements.txt
asgiref==3.8.1
autopep8==2.3.2
black==25.1.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
click==8.1.8
colorama==0.4.6
cssselect2==0.8.0
diff-match-patch==20241021
dj-database-url==2.2.0
Django==4.2.23
django-ckeditor==6.7.2
django-debug-toolbar==5.1.0
django-filter==25.1
django-import-export==4.3.7
django-js-asset==2.2.0
django-modeltranslation==0.19.13
django-simple-history==3.8.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
Faker==37.0.2
ffs==0.0.8.2
flake8==7.2.0
fonttools==4.58.4
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isort==6.0.1
Jinja2==3.1.6
letter==0.5
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
openpyxl==3.1.5
packaging==25.0
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pydyf==0.11.0
pyflakes==3.3.2
pyphen==0.17.2
reportlab==4.3.1
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
tzdata==2025.1
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1